Create initial obj-c wrapper class
